+++ templates/security/register.html.twig
@@ -1,0 +1,17 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Wild Series | Inscription{% endblock %}
+
+{% block body %}
+
+    <div class="container">
+        <h1 class="mb-5 mt-5">Inscription</h1>
+        {{ form_start(registrationForm) }}
+        {{ form_row(registrationForm.email) }}
+        {{ form_row(registrationForm.plainPassword) }}
+        {{ form_row(registrationForm.agreeTerms) }}
+
+        <button class="btn btn-lg round mt-5">S'inscrire</button>
+        {{ form_end(registrationForm) }}
+    </div>
+{% endblock %}
